issue #24 , added video sources
Video sources are now able to be created, when creating a video source with a youtube url, it turns it to an embed and plays it. If the link is not a youtube link it plays it as is. Additionaly, moving between pdfs and videos is working, changing between viewers depending on the source type. - Saul Robles