The project page shows the user any projects created by them and starting of a project viewer page
When the user loads on the project page all the projects are fetch and shown for the user.
If the user clicks on a project they are routed to a project viewer page with direct routing. This supports the user stories of users creating projects and viewing them and adding/Removing sources to the project.

-Corey Williams