fixed project page not updating with new project, project details are added on the viewer, and users can delete projects
When a project is created on the project page now the UI will dynamically update displaying the new project.
When viewing a project on the project viewer the title and description are now visisble on the page and the User can now delete a project on the project viewer Page, this supports the user story of users creating and removing projects.

-Corey Williams